Remove pelican-alias plugin
This plugin was used after I switched from Jekyll to Pelican to make
sure old URLs don't break. Because it hasn't been maintained for a
while, I decided to just manually recreate the redirects it set up for
the two most recent blog posts, which are also the ones that have been
cited by other publications.